Replace TF2 game code with official Valve Source SDK 2013.

+++ game/server/ai_behavior_follow.cpp
@@ -53,7 +53,7 @@ struct AI_Follower_t
 	}
 
 	AIHANDLE 			hFollower;
-	intp					slot;
+	int					slot;
 	AI_FollowNavInfo_t	navInfo;
 	AI_FollowGroup_t *	pGroup;	// backpointer for efficiency
 };
@@ -2569,8 +2569,9 @@ bool CAI_FollowManager::AddFollower( CBaseEntity *pTarget, CAI_BaseNPC *pFollowe
 		iterNode->pGroup		= pGroup;
 
 		pGroup->slotUsage.Set( slot );
+		
 		CalculateFieldsFromSlot( pSlot, &iterNode->navInfo );
-
+		
 		pHandle->m_hFollower = i;
 		pHandle->m_pGroup = pGroup;
 		return true;
@@ -2643,7 +2644,7 @@ bool CAI_FollowManager::RedistributeSlots( AI_FollowGroup_t *pGroup )
 		intp  h			= pGroup->followers.Head();
 		intp  hBest 		= pGroup->followers.InvalidIndex();
 		float 			   distSqBest = FLT_MAX;
-
+		
 		while ( h != pGroup->followers.InvalidIndex() )
 		{
 			AI_Follower_t *p = &pGroup->followers[h];
@@ -2737,7 +2738,7 @@ void CAI_FollowManager::RemoveFollower( AI_FollowManagerInfoHandle_t& hInfo )
 		AI_FollowGroup_t *pGroup = hInfo.m_pGroup;
 		AI_Follower_t* iterNode = &pGroup->followers[hInfo.m_hFollower];
 
-		intp slot = iterNode->slot;
+		int slot = iterNode->slot;
 		pGroup->slotUsage.Clear( slot );
 		pGroup->followers.Remove( hInfo.m_hFollower );
 		if ( pGroup->followers.Count() == 0 )
